refactor: unify RMSNormGated and FusedRMSNormGated
